aries_design: created storybook.dart
