+++ packages/aries_designs/lib/aries_designs.dart
@@ -1,13 +1,1 @@
-import 'package:flutter/material.dart';
-import 'package:storybook_flutter/storybook_flutter.dart';
 
-class AriesDesignStoryBook extends StatelessWidget {
-  const AriesDesignStoryBook({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Storybook(
-      stories: const [],
-    );
-  }
-}
