+++ packages/aries_designs/lib/src/storybook/storybook.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:storybook_flutter/storybook_flutter.dart';
+
+class AriesDesignStoryBook extends StatelessWidget {
+  const AriesDesignStoryBook({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Storybook(
+      stories: const [],
+    );
+  }
+}
